pyproject.toml
[project]
name = "ibovespa-backend"
version = "0.1.0"
description = ""
authors = [
    {name = "Caio",email = "caiomorozini@outlook.com.br"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard] (>=0.117.1,<0.118.0)",
    "pydantic-settings (>=2.11.0,<3.0.0)",
    "passlib[bcrypt] (>=1.7.4,<2.0.0)",
    "sqlalchemy (>=2.0.43,<3.0.0)",
    "databases (>=0.9.0,<0.10.0)",
    "python-jose (>=3.5.0,<4.0.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "bcrypt (>=5.0.0,<6.0.0)",
    "argon2-cffi (>=25.1.0,<26.0.0)",
    "pandas (>=2.3.3,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
